+++ src/flag.rs
@@ -5,19 +5,26 @@ pub struct Flags {
     timestamp:    bool,
     append:       bool,
     instead_crlf: bool,
+    hexdump:      bool,
     debug:        bool,
     write_file:   Option<String>,
 }
 
 impl Flags {
-    pub fn new(nocolor: bool, timestamp: bool, append: bool, instead_crlf: bool, debug: bool,
-        write_file: Option<String>) -> Self
-    {
+    pub fn new(nocolor:      bool,
+               timestamp:    bool,
+               append:       bool,
+               instead_crlf: bool,
+               hexdump:      bool,
+               debug:        bool,
+               write_file:   Option<String>,
+    ) -> Self {
         Self {
             nocolor,
             timestamp,
             append,
             instead_crlf,
+            hexdump,
             debug,
             write_file,
         }
@@ -31,6 +38,8 @@ impl Flags {
     pub fn append_mut(&mut self)       -> &mut bool { &mut self.append }
     pub fn instead_crlf(&self)         -> &bool     { &self.instead_crlf }
     pub fn instead_crlf_mut(&mut self) -> &mut bool { &mut self.instead_crlf }
+    pub fn hexdump(&self)              -> &bool     { &self.hexdump }
+    pub fn hexdump_mut(&mut self)      -> &mut bool { &mut self.hexdump }
     pub fn debug(&self)                -> &bool     { &self.debug }
     pub fn debug_mut(&mut self)        -> &mut bool { &mut self.debug }