+++ src/lib.rs
@@ -1,12 +1,13 @@
 pub mod color;
-pub mod config;
 pub mod default;
 pub mod file_read;
 pub mod flag;
 pub mod getch;
+pub mod hexdump;
 pub mod negotiation;
 pub mod repl;
 pub mod serial;
 pub mod setting;
+pub mod ssh;
 pub mod tcp;
 pub mod telnet;