+++ src/serial.rs
@@ -1,24 +1,25 @@
-use std::thread;
 use std::path::Path;
 use std::sync::{mpsc, Arc, Mutex};
+use std::time::Duration;
 
 use crate::repl;
 use crate::flag;
 use crate::setting;
 use crate::getch::{Getch, Key};
 
-use serialport::SerialPortSettings;
-
-
-pub fn run(port_name: String,
-           settings:  SerialPortSettings,
-           mut flags: flag::Flags,
-           params:    Option<setting::Params>)
-{
-    let receiver = serialport::open_with_settings(&port_name, &settings).unwrap_or_else(|e| {
-        eprintln!("Failed to open \"{}\". Error: {}", port_name, e);
-        std::process::exit(1);
-    });
+pub async fn run(
+    port_name: String,
+    baud_rate: u32,
+    mut flags: flag::Flags,
+    params:    Option<setting::Params>,
+){
+    let receiver = serialport::new(&port_name, baud_rate)
+        .timeout(Duration::from_millis(10))
+        .open()
+        .unwrap_or_else(|e| {
+            eprintln!("Failed to open \"{}\". Error: {}", port_name, e);
+            std::process::exit(1);
+        });
     let transmitter = receiver.try_clone().expect("Failed to clone from receiver");
 
     let (tx, rx) = mpsc::channel();
@@ -58,13 +59,11 @@ pub fn run(port_name: String,
     let flags       = Arc::new(Mutex::new(flags));
     let flags_clone = flags.clone();
 
-    // Receiver
-    let handle = thread::spawn(move || {
-        repl::receiver(receiver, rx, flags_clone, params);
-    });
-
-    // Transmitter
-    repl::transmitter(transmitter, tx, flags);
-
-    handle.join().unwrap();
+    tokio::select! {
+        _ = tokio::spawn(repl::receiver(receiver, rx, flags_clone, params)) => {
+            println!("\n\x1b[0mDisconnected.");
+            std::process::exit(0);
+        }
+        _ = tokio::spawn(repl::transmitter(transmitter, tx, flags)) => {}
+    }
 }